refactor (Position) : 패키지 분리
