+++ src/main/java/chess/position/File.java
@@ -1,4 +1,4 @@
-package chess.board;
+package chess.position;
 
 import java.util.Arrays;
