+++ src/test/java/chess/piece/BishopTest.java
@@ -3,8 +3,8 @@ package chess.piece;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import chess.board.Direction;
-import chess.board.Position;
+import chess.position.Direction;
+import chess.position.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
