+++ src/test/java/chess/piece/KnightTest.java
@@ -2,8 +2,8 @@ package chess.piece;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import chess.board.Direction;
-import chess.board.Position;
+import chess.position.Direction;
+import chess.position.Position;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
