+++ src/test/java/chess/position/RankTest.java
@@ -1,4 +1,4 @@
-package chess.board;
+package chess.position;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
